requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
attrs==24.1.0
bitarray==2.9.2
cached-property==1.5.2
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==1.0.2
cytoolz==0.12.3
eth-account==0.11.2
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-tester==0.11.0b2
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
hexbytes==0.3.1
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
parsimonious==0.10.0
protobuf==5.27.3
py-ecc==7.0.1
py-evm==0.10.1b1
py-geth==4.4.0
pycryptodome==3.20.0
pyunormalize==15.1.0
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
rpds-py==0.19.1
safe-pysha3==1.0.4
semantic-version==2.10.0
sortedcontainers==2.4.0
toolz==0.12.1
trie==3.0.1
typing_extensions==4.12.2
urllib3==2.2.2
web3==6.20.1
websockets==12.0
yarl==1.9.4
